+++ R/learner_RWeka_classif_PART.R
@@ -77,38 +77,15 @@ LearnerClassifPART = R6Class("LearnerClassifPART",
 
   private = list(
     .train = function(task) {
-      params = self$param_set$get_values(tags = "train")
-      ctrl_arg_names = weka_control_args(RWeka::PART)
-      arg_names = setdiff(names(params), ctrl_arg_names)
-      ctrl = params[which(names(params) %in% ctrl_arg_names)]
-      pars = params[which(names(params) %nin% ctrl_arg_names)]
-
-      if (length(ctrl) > 0L) {
-        names(ctrl) = gsub("_", replacement = "-", x = names(ctrl))
-        ctrl = invoke(RWeka::Weka_control, .args = ctrl)
-      }
-
-      f = task$formula()
-      data = task$data()
-      invoke(RWeka::PART, formula = f, data = data, control = ctrl, .args = pars)
+      weka_learner = RWeka::PART
+      pars = self$param_set$get_values(tags = "train")
+      rweka_train(task$data(), task$formula(), pars, weka_learner)
     },
 
     .predict = function(task) {
-      response = NULL
-      prob = NULL
-      newdata = ordered_features(task, self)
       pars = self$param_set$get_values(tags = "predict")
-
-      if (self$predict_type == "response") {
-        response = invoke(predict, self$model, newdata = newdata, type = "class",
-          .args = pars
-        )
-      } else {
-        prob = invoke(predict, self$model, newdata = newdata, type = "prob",
-          .args = pars
-        )
-      }
-      list(response = response, prob = prob)
+      newdata = ordered_features(task, self)
+      rweka_predict(newdata, pars, self$predict_type, self$model)
     }
   )
 )