+++ R/learner_RWeka_regr_decision_stump.R
@@ -1,0 +1,78 @@
+#' @title Regression Decision Stump Learner
+#' @author damirpolat
+#' @name mlr_learners_regr.decision_stump
+#'
+#' @description
+#' Decision Stump Learner.
+#' Calls [RWeka::DecisionStump()] from \CRANpkg{RWeka}.
+#'
+#' @section Custom mlr3 parameters:
+#' - `output_debug_info`:
+#'   - original id: output-debug-info
+#'
+#' - `do_not_check_capabilities`:
+#'   - original id: do-not-check-capabilities
+#'
+#' - `num_decimal_places`:
+#'   - original id: num-decimal-places
+#'
+#' - `batch_size`:
+#'   - original id: batch-size
+#'
+#' - Reason for change: This learner contains changed ids of the following control arguments
+#' since their ids contain irregular pattern
+#'
+#'
+#' @templateVar id regr.decision_stump
+#' @template learner
+#'
+#' @template seealso_learner
+#' @template example
+#' @export
+LearnerRegrDecisionStump = R6Class("LearnerRegrDecisionStump",
+  inherit = LearnerRegr,
+  public = list(
+    #' @description
+    #' Creates a new instance of this [R6][R6::R6Class] class.
+    initialize = function() {
+      param_set = ps(
+        subset = p_uty(tags = "train"),
+        na.action = p_uty(tags = "train"),
+        output_debug_info = p_lgl(default = FALSE, tags = "train"),
+        do_not_check_capabilities = p_lgl(default = FALSE,
+                                          tags = "train"),
+        num_decimal_places = p_int(default = 2L, lower = 1L,
+                                   tags = "train"),
+        batch_size = p_int(default = 100L, lower = 1L, tags = "train"),
+        options = p_uty(default = NULL, tags = "train")
+      )
+
+      super$initialize(
+        id = "regr.decision_stump",
+        packages = "RWeka",
+        feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
+        predict_types = "response",
+        param_set = param_set,
+        properties = "missings",
+        man = "mlr3extralearners::mlr_learners_regr.decision_stump",
+        label = "Decision Stump"
+      )
+    }
+  ),
+
+  private = list(
+    .train = function(task) {
+      weka_learner = RWeka::DecisionStump
+      pars = self$param_set$get_values(tags = "train")
+      rweka_train(task$data(), task$formula(), pars, weka_learner)
+    },
+
+    .predict = function(task) {
+      pars = self$param_set$get_values(tags = "predict")
+      newdata = ordered_features(task, self)
+      rweka_predict(newdata, pars, self$predict_type, self$model)
+    }
+  )
+)
+
+.extralrns_dict$add("regr.decision_stump", LearnerRegrDecisionStump)